datadog_protos

Module agent

Agent definitions.

Modules

config_event

Nested message and enum types in ConfigEvent.

generate_container_id_from_origin_info_request

Nested message and enum types in GenerateContainerIDFromOriginInfoRequest.

get_telemetry_response

Nested message and enum types in GetTelemetryResponse.

Structs

AdvancedAdIdentifier

AgentClient

The greeting service definition.

AgentSecureClient

AutodiscoveryStreamResponse

CaptureTriggerRequest

Dogstatsd capture types

CaptureTriggerResponse

Config

ConfigEvent

ConfigSetting

ConfigSnapshot

ConfigStreamRequest

ConfigUpdate

Container

ContainerAllocatedResource

ContainerImage

ContainerPort

ContainerResources

ContainerState

DeprecatedFilter

EcsTask

Entity

EntityId

EntityMeta

FetchEntityRequest

FetchEntityResponse

GenerateContainerIdFromOriginInfoRequest

GenerateContainerIdFromOriginInfoResponse

GetFlareFilesRequest

GetFlareFilesResponse

GetStatusDetailsRequest

GetStatusDetailsResponse

GetTelemetryRequest

GetTelemetryResponse

HostTagReply

HostTagRequest

HostnameReply

The response message containing the requested hostname

HostnameRequest

KubeEndpointsIdentifier

KubeNamespacedName

KubernetesPod

KubernetesPodOwner

OrchestratorContainer

RegisterRemoteAgentRequest

RegisterRemoteAgentResponse

RemoteAgentServer

Service exposed by remote agents to allow querying by the Core Agent.

StatusSection

StreamTagsEvent

StreamTagsRequest

StreamTagsResponse

TaggerState

TaggerStateResponse

UnixDogstatsdMsg

UDS Capture The message contains the payload and the ancillary info

WorkloadmetaEntityId

WorkloadmetaEvent

WorkloadmetaFilter

WorkloadmetaStreamRequest

WorkloadmetaStreamResponse

Enums

ConfigEventType

ContainerHealth

ContainerStatus

EcsLaunchType

EventType

Runtime

TagCardinality

WorkloadmetaEventType

WorkloadmetaKind

WorkloadmetaSource

Traits

RemoteAgent

Generated trait containing gRPC methods that should be implemented for use with RemoteAgentServer.