datadog_protos::agent

Struct AgentSecureClient

pub struct AgentSecureClient<T> { /* private fields */ }

Implementations

impl<T> AgentSecureClient<T>
where T: GrpcService<Body>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> AgentSecureClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<Body>, Response = Response<<T as GrpcService<Body>>::ResponseBody>>, <T as Service<Request<Body>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn tagger_stream_entities( &mut self, request: impl IntoRequest<StreamTagsRequest>, ) -> Result<Response<Streaming<StreamTagsResponse>>, Status>

subscribes to added, removed, or changed entities in the Tagger and streams them to clients as events.

pub async fn tagger_generate_container_id_from_origin_info( &mut self, request: impl IntoRequest<GenerateContainerIdFromOriginInfoRequest>, ) -> Result<Response<GenerateContainerIdFromOriginInfoResponse>, Status>

Generates a container ID from Origin Info.

pub async fn tagger_fetch_entity( &mut self, request: impl IntoRequest<FetchEntityRequest>, ) -> Result<Response<FetchEntityResponse>, Status>

fetches an entity from the Tagger with the desired cardinality tags.

pub async fn dogstatsd_capture_trigger( &mut self, request: impl IntoRequest<CaptureTriggerRequest>, ) -> Result<Response<CaptureTriggerResponse>, Status>

Trigger a dogstatsd capture. Only one capture can be triggered at a time.

pub async fn dogstatsd_set_tagger_state( &mut self, request: impl IntoRequest<TaggerState>, ) -> Result<Response<TaggerStateResponse>, Status>

Set the tagger state for dogstatsd.

pub async fn client_get_configs( &mut self, request: impl IntoRequest<ClientGetConfigsRequest>, ) -> Result<Response<ClientGetConfigsResponse>, Status>

pub async fn get_config_state( &mut self, request: impl IntoRequest<()>, ) -> Result<Response<GetStateConfigResponse>, Status>

pub async fn client_get_configs_ha( &mut self, request: impl IntoRequest<ClientGetConfigsRequest>, ) -> Result<Response<ClientGetConfigsResponse>, Status>

pub async fn get_config_state_ha( &mut self, request: impl IntoRequest<()>, ) -> Result<Response<GetStateConfigResponse>, Status>

pub async fn create_config_subscription( &mut self, request: impl IntoStreamingRequest<Message = ConfigSubscriptionRequest>, ) -> Result<Response<Streaming<ConfigSubscriptionResponse>>, Status>

pub async fn reset_config_state( &mut self, request: impl IntoRequest<()>, ) -> Result<Response<ResetStateConfigResponse>, Status>

pub async fn workloadmeta_stream_entities( &mut self, request: impl IntoRequest<WorkloadmetaStreamRequest>, ) -> Result<Response<Streaming<WorkloadmetaStreamResponse>>, Status>

Subscribes to added, removed, or changed entities in the Workloadmeta and streams them to clients as events.

pub async fn register_remote_agent( &mut self, request: impl IntoRequest<RegisterRemoteAgentRequest>, ) -> Result<Response<RegisterRemoteAgentResponse>, Status>

Registers a remote agent.

pub async fn refresh_remote_agent( &mut self, request: impl IntoRequest<RefreshRemoteAgentRequest>, ) -> Result<Response<RefreshRemoteAgentResponse>, Status>

Refresh a remote agent.

pub async fn autodiscovery_stream_config( &mut self, request: impl IntoRequest<()>, ) -> Result<Response<Streaming<AutodiscoveryStreamResponse>>, Status>

Subscribes to autodiscovery config updates

pub async fn get_host_tags( &mut self, request: impl IntoRequest<HostTagRequest>, ) -> Result<Response<HostTagReply>, Status>

Get the host tags

pub async fn stream_config_events( &mut self, request: impl IntoRequest<ConfigStreamRequest>, ) -> Result<Response<Streaming<ConfigEvent>>, Status>

Streams config events to the remote agent.