datadog_protos::agent

Struct WorkloadmetaFilter

pub struct WorkloadmetaFilter {
    pub kinds: Vec<i32>,
    pub source: i32,
    pub event_type: i32,
}

Fields
kinds: Vec<i32>source: i32event_type: i32
Implementations

impl WorkloadmetaFilter

pub fn kinds( &self, ) -> FilterMap<Cloned<Iter<'_, i32>>, fn(i32) -> Option<WorkloadmetaKind>>

Returns an iterator which yields the valid enum values contained in kinds.

pub fn push_kinds(&mut self, value: WorkloadmetaKind)

Appends the provided enum value to kinds.

pub fn source(&self) -> WorkloadmetaSource

Returns the enum value of source, or the default if the field is set to an invalid enum value.

pub fn set_source(&mut self, value: WorkloadmetaSource)

Sets source to the provided enum value.

pub fn event_type(&self) -> WorkloadmetaEventType

Returns the enum value of event_type, or the default if the field is set to an invalid enum value.

pub fn set_event_type(&mut self, value: WorkloadmetaEventType)

Sets event_type to the provided enum value.