datadog_protos::agent

Trait RemoteAgent

pub trait RemoteAgent:
    Send
    + Sync
    + 'static {
    // Required methods
    fn get_status_details<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetStatusDetailsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetStatusDetailsResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_flare_files<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetFlareFilesRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetFlareFilesResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_telemetry<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetTelemetryRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetTelemetryResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with RemoteAgentServer.

Required Methods

fn get_status_details<'life0, 'async_trait>( &'life0 self, request: Request<GetStatusDetailsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetStatusDetailsResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets the status details of a remote agent.

fn get_flare_files<'life0, 'async_trait>( &'life0 self, request: Request<GetFlareFilesRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetFlareFilesResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets all relevant flare files of a remote agent.

fn get_telemetry<'life0, 'async_trait>( &'life0 self, request: Request<GetTelemetryRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetTelemetryResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets the telemetry data of a remote agent.