datadog_protos::agent::v1::remote_agent_server

Trait RemoteAgent

pub trait RemoteAgent:
    Send
    + Sync
    + 'static {
    // Required methods
    fn register_remote_agent<'life0, 'async_trait>(
        &'life0 self,
        request: Request<RegisterRemoteAgentRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<RegisterRemoteAgentResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn refresh_remote_agent<'life0, 'async_trait>(
        &'life0 self,
        request: Request<RefreshRemoteAgentRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<RefreshRemoteAgentResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn report_remote_agent_event<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ReportRemoteAgentEventRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ReportRemoteAgentEventResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with RemoteAgentServer.

Required Methods

fn register_remote_agent<'life0, 'async_trait>( &'life0 self, request: Request<RegisterRemoteAgentRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<RegisterRemoteAgentResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Registers a remote agent.

fn refresh_remote_agent<'life0, 'async_trait>( &'life0 self, request: Request<RefreshRemoteAgentRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<RefreshRemoteAgentResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Refreshes a remote agent session.

fn report_remote_agent_event<'life0, 'async_trait>( &'life0 self, request: Request<ReportRemoteAgentEventRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ReportRemoteAgentEventResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Reports one or more operational events from a remote agent.