datadog_protos

Module remote_config

Remote Configuration definitions.

Modules

config_subscription_request

Nested message and enum types in ConfigSubscriptionRequest.

config_subscription_state

Nested message and enum types in ConfigSubscriptionState.

Structs

Client

ClientAgent

ClientGetConfigsRequest

ClientGetConfigsResponse

ClientState

ClientTracer

ClientUpdater

ConfigMetas

ConfigState

ConfigSubscriptionRequest

ConfigSubscriptionRequest is used to manage the state of the stream created using CreateConfigSubscription.

ConfigSubscriptionResponse

ConfigSubscriptionResponse is streamed from CreateConfigSubscription with updates for matching clients and products.

ConfigSubscriptionState

ConfigSubscriptionState describes the state of a config subscription.

DelegatedMeta

DirectorMetas

File

FileMetaState

GetStateConfigResponse

LatestConfigsRequest

LatestConfigsResponse

OrgDataResponse

OrgStatusResponse

PackageState

PackageStateTask

ResetStateConfigResponse

TargetFileHash

TargetFileMeta

TaskError

TopMeta

TracerPredicateV1

TracerPredicates

Enums

ConfigStatus

ConfigSubscriptionProducts

ConfigSubscriptionProducts is used to targets specific products for tracking with a ConfigSubscriptionRequest.

TaskState